Renaming namespace ViperFpga to Fpga to reflect change in folder name
