+++ src/ArduinoViperFpga.cpp
@@ -19,7 +19,7 @@ extern void enableFpgaClock(); /* Defined within MKRVIDOR4000/variant.cpp */
  * CTOR/DTOR
  **************************************************************************************/
 
-ArduinoViperFpga::ArduinoViperFpga(ViperFpga::SpiSelectFunc select, ViperFpga::SpiDeselectFunc deselect, ViperFpga::SpiTransferFunc transfer)
+ArduinoViperFpga::ArduinoViperFpga(Fpga::SpiSelectFunc select, Fpga::SpiDeselectFunc deselect, Fpga::SpiTransferFunc transfer)
 : _io(select, deselect, transfer)
 {
 
@@ -37,23 +37,23 @@ ArduinoViperFpga::Status ArduinoViperFpga::begin()
 
 uint8_t ArduinoViperFpga::getRevNum()
 {
-  return _io.read(ViperFpga::Register::FPGA_REV_NUM);
+  return _io.read(Fpga::Register::FPGA_REV_NUM);
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityRed(uint8_t const red)
 {
-  _io.write(ViperFpga::Register::RGB_LED_RED_INTENSITY, red);
+  _io.write(Fpga::Register::RGB_LED_RED_INTENSITY, red);
   return Status::OK;
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityGreen(uint8_t const green)
 {
-  _io.write(ViperFpga::Register::RGB_LED_GREEN_INTENSITY, green);
+  _io.write(Fpga::Register::RGB_LED_GREEN_INTENSITY, green);
   return Status::OK;
 }
 
 ArduinoViperFpga::Status ArduinoViperFpga::setIntensityBlue(uint8_t const blue)
 {
-  _io.write(ViperFpga::Register::RGB_LED_BLUE_INTENSITY, blue);
+  _io.write(Fpga::Register::RGB_LED_BLUE_INTENSITY, blue);
   return Status::OK;
 }
