+++ src/Fpga/FpgaConst.h
@@ -13,7 +13,7 @@
  * NAMESPACE
  **************************************************************************************/
 
-namespace ViperFpga
+namespace Fpga
 {
 
 /**************************************************************************************
@@ -32,4 +32,4 @@ enum class Register : uint8_t
  * NAMESPACE
  **************************************************************************************/
 
-} /* namespace ViperFpga */
+} /* namespace Fpga */
