+++ src/Fpga/FpgaIo.cpp
@@ -15,7 +15,7 @@
  * NAMESPACE
  **************************************************************************************/
 
-namespace ViperFpga
+namespace Fpga
 {
 
 /**************************************************************************************
@@ -60,4 +60,4 @@ void FpgaIo::write(Register const reg, uint8_t const reg_val)
  * NAMESPACE
  **************************************************************************************/
 
-} /* namespace ViperFpga */
+} /* namespace Fpga */
