+++ src/Fpga/FpgaIo.h
@@ -22,7 +22,7 @@
  * NAMESPACE
  **************************************************************************************/
 
-namespace ViperFpga
+namespace Fpga
 {
 
 /**************************************************************************************
@@ -61,6 +61,6 @@ private:
  * NAMESPACE
  **************************************************************************************/
 
-} /* namespace ViperFpga */
+} /* namespace Fpga */
 
 #endif /* ARDUINO_VIPER_FPGA_IO_H_ */
